requirements.txt
python-chess
numpy
pandas
matplotlib
jupyter